Moved etl related functions out of cli to etl
